+++ h2integrate/converters/iron/humbert_ewin_perf.py
@@ -1,3 +1,21 @@
+"""Iron electronwinning performance model based on Humbert et al.
+
+This module contains H2I performance configs and components for modeling iron electrowinning. It is
+based of the work of Humbert et al. (doi.org/10.1007/s40831-024-00878-3) which reviews performance
+and TEA literature for three different types of iron electrowinning:
+    - Aqueous Hydroxide Electrolysis (AHE)
+    - Molten Salt Electrolysis (MSE)
+    - Molten Oxide Electrolysis (MOE)
+
+This technology is selected in the tech_config as the performance_model.
+"humbert_electrowinning_performance"
+
+Classes:
+    HumbertEwinConfig: Sets the required model_inputs fields.
+    HumbertEwinPerformanceComponent: Defines initialize(), setup(), and compute() methods.
+
+"""
+
 import numpy as np
 import openmdao.api as om
 from attrs import field, define
@@ -8,6 +26,18 @@ from h2integrate.core.validators import contains
 
 @define
 class HumbertEwinConfig(BaseConfig):
+    """Configuration class for the Humbert iron electrowinning performance model.
+
+    Args:
+        electrolysis_type (str): The type of electrowinning being performed. Options:
+            "ahe": Aqueous Hydroxide Electrolysis (AHE)
+            "mse": Molten Salt Electrolysis (MSE)
+            "moe": Molten Oxide Electrolysis (MOE)
+        ore_fe_wt_pct (float): The iron content of the ore coming in, expressed as a percentage.
+        capacity_mw (float): The MW electrical capacity of the electrowinning plant.
+
+    """
+
     electrolysis_type: str = field(
         kw_only=True, converter=(str.lower, str.strip), validator=contains(["ahe", "mse", "moe"])
     )  # product selection
@@ -16,8 +46,26 @@ class HumbertEwinConfig(BaseConfig):
 
 
 class HumbertEwinPerformanceComponent(om.ExplicitComponent):
-    """
-    Humbert: doi.org/10.1007/s40831-024-00878-3
+    """OpenMDAO component for the Humbert iron electrowinning performance model.
+
+    Inputs:
+        electricity_in (array): Electric power input available in kW for each timestep.
+        iron_ore_in (array): Iron ore mass flow available in kg/h for each timestep.
+        ore_fe_wt_pct (float): The iron content of the ore coming in, expressed as a percentage.
+        spec_energy_cons_fe (float): The specific electrical energy consumption required to win
+            pure iron (Fe) from iron ore. These are currently calculated as averages between the
+            high and low stated values in Table 10 of Humbert et al., but this is exposed as an
+            OpenMDAO variable to probe the effect of specific energy consumption on iron cost.
+        capacity (float): The electrical capacity of the electrowinning plant in MW.
+
+    Outputs:
+        electricity_consumed (array): Electric power consumption in kW for each timestep.
+        limiting_input (array): An array of integers indicating which input is the limiting factor
+            for iron production at each timestep: 0 = iron ore, 1 = electricity, 2 = capacity
+        sponge_iron_out (array): Sponge iron production in kg/h for each timestep.
+        total_sponge_iron_produced (float): Total annual sponge iron production in kg/y.
+        output_capacity (float): Maximum possible annual sponge iron production in kg/y.
+
     """
 
     def initialize(self):