+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Updates models for NumPy version 2.4.0
 - Update test values for WOMBAT update to 0.13.0
 - Added standlone iron DRI and steel EAF performance and cost models
+- Added iron electrowinning model
 - Added capability to have transport models that require user input parameters
 - Add geologic hydrogen surface processing converter
 - Add baseclass for caching functionality